% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{anno_cgc_cna}
\alias{anno_cgc_cna}
\title{Annotate CNA with cgc_gr IDs}
\usage{
anno_cgc_cna(gr, cgc_gr, which_genome)
}
\arguments{
\item{gr}{GRanges object}

\item{cgc_gr}{cancer gene census GRanges}

\item{which_genome}{the genome assembly used, "hg19" or "hg38"}
}
\value{
GRanges object with annotation
}
\description{
Annotate CNA with cgc_gr IDs
}
