% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{at_least_two}
\alias{at_least_two}
\title{Find consensus of at least two callers using GRanges 'superset'}
\usage{
at_least_two(var_list, gr_super, tag)
}
\arguments{
\item{var_list}{is a nested list of [[caller]][[samples1..n]]}

\item{gr_super}{is a GRanges superset from gr_super_set()}
}
\value{
GRanges object of consensus per sample
}
\description{
Find consensus of at least two callers using GRanges 'superset'
}
