% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{bin_granges}
\alias{bin_granges}
\title{Take input of a GRanges object, and use that seqinfo to bin by bin_size}
\usage{
bin_granges(
  gr,
  genome = NULL,
  bin_size = NULL,
  mcol_in,
  mcol_in_lim = NULL,
  one_to_x = TRUE
)
}
\arguments{
\item{gr}{GRanges object with seqinfo used to make bins}

\item{genome}{use this to add seqinfo if none extant (e.g. "hg19", "hg38")}

\item{bin_size}{base pairs by which to bin}

\item{mcol_in}{string naming mcols to use for binning (one only)}

\item{mcol_in_lim}{numeric on which to screen mcol_in column; if length>1, take all outside the range suplied}

\item{one_to_x}{logical indicating whether to use chr1..chrX for bin, as opposed to what is in input GRanges}
}
\value{
GRanges binned on mcol_in
}
\description{
Take input of a GRanges object, and use that seqinfo to bin by bin_size
}
