% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_pairtree.R
\name{cna_master_muts}
\alias{cna_master_muts}
\title{Parse CN file and output as list}
\usage{
cna_master_muts(cn_master, mut_df_filt)
}
\arguments{
\item{cn_master}{is a pattern that matches CN input filenames
e.g. *.ENS.facets.CNA.master.tsv from fctcons}

\item{mut_df_filt}{object with mut_df_filt from above}
}
\value{
cn values overlapping mut_df_filt loci
}
\description{
Parse CN file and output as list
}
