% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{combn_at_least_two}
\alias{combn_at_least_two}
\title{Find consensus of at least two callers using GRanges superset}
\usage{
combn_at_least_two(var_list, gr_super, callers, samp, tag)
}
\arguments{
\item{var_list}{is a nested list of [[caller]][[samples1..n]]}

\item{callers}{are the named variant callers}

\item{samp}{is the sample being operated on}

\item{tag}{is a string to tag output files}
}
\value{
GRanges object of consensus per sample
}
\description{
Find consensus of at least two callers using GRanges superset
}
