% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_gridss.R
\name{download_chimerkb4}
\alias{download_chimerkb4}
\title{Download and parse ChimerKB v4 for annotating known fusions}
\usage{
download_chimerkb4()
}
\value{
chimerkb tibble of parsed ChimerKB v4
}
\description{
Download and parse ChimerKB v4 for annotating known fusions
}
