% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{facets_cna_call}
\alias{facets_cna_call}
\title{Facets functions}
\usage{
facets_cna_call(input_csv)
}
\arguments{
\item{input_csv}{path to CSV format file as per requirements of Facets}
}
\value{
none, saves plots and writes output in current dir
}
\description{
Standard Facets CNA call
}
