% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{facets_jointsegs_parse_to_gr}
\alias{facets_jointsegs_parse_to_gr}
\title{Parse jointsegs into GRanges object}
\usage{
facets_jointsegs_parse_to_gr(
  jointsegs_in,
  which_genome,
  anno = NULL,
  cgc_gr = NULL,
  bsgenome = NULL
)
}
\arguments{
\item{jointsegs_in}{filename (output from Facets) with columns:
"seg", "num.mark", "nhet", "cnlr.median", "mafR", "segclust", "cnlr.median.clust", "mafR.clust", "cf.em", "tcn.em", "lcn.em"}

\item{which_genome}{the genome assembly used, "hg19" or "hg38"}

\item{anno}{annotation strategy, "ENS" or "CGC"}

\item{cgc_gr}{GRanges object of cancer gene census from COSMIC}

\item{bsgenome}{which version of bsgenome to use}
}
\value{
a GRanges object with annotated jointsegs from input
}
\description{
Parse jointsegs into GRanges object
}
