% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{gr_super_alt_plot}
\alias{gr_super_alt_plot}
\title{Wrapper of main functions}
\usage{
gr_super_alt_plot(
  var_list,
  name_callers,
  impacts,
  taga,
  included_order,
  which_genome
)
}
\arguments{
\item{var_list}{is a nested list of VEP annotated [[caller]][[samples1..n]]}

\item{name_callers}{two of the variant callers}

\item{impacts}{VEP impacts (one or combination of "HIGH", "MODERATE", "MODIFIER", "LOW")}

\item{included_order}{oredering of samples for plotting}

\item{which_genome}{hg19 or hg38}

\item{raw_list}{is a nested list of unfiltered calls [[caller]][[samples1..n]]}

\item{tag}{is a string used to tag output}
}
\value{
GRanges object of all  of single-letter HGVS protein IDs
}
\description{
Wrapper of main functions
}
