% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{gr_super_set}
\alias{gr_super_set}
\title{Create GRanges 'superset' from GRanges per-sample, per-caller variants
   contains all variants found per sample in any caller}
\usage{
gr_super_set(var_list, name_callers, impacts)
}
\arguments{
\item{var_list}{is a nested list of [[caller]][[samples1..n]]}

\item{name_callers}{is a set of two callers used for initial screening}

\item{impacts}{VEP impacts (one or combination of "HIGH", "MODERATE", "MODIFIER", "LOW")}
}
\value{
GRanges 'superset' of all callers, and samples therein
}
\description{
Create GRanges 'superset' from GRanges per-sample, per-caller variants
   contains all variants found per sample in any caller
}
