% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_gridss.R
\name{gridss_annotate_gr}
\alias{gridss_annotate_gr}
\title{Annotate a GRanges object from GRIDSS}
\usage{
gridss_annotate_gr(gr, which_genome)
}
\arguments{
\item{gr}{GRanges object}

\item{which_genome, }{genome assembly used ("hg19", "hg38")}
}
\value{
annotated GRanges object
}
\description{
Annotate a GRanges object from GRIDSS
}
