% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_gridss.R
\name{gridss_bedpe_list}
\alias{gridss_bedpe_list}
\title{Produce a list of bedpe, from a list of breakpoint GRanges,
   with score > qual_filt and both breakpoints within annotated genes}
\usage{
gridss_bedpe_list(bpgr_list, which_genome, qual_filt = 2)
}
\arguments{
\item{bpgr_list}{list of per-sample breakpont GRanges objects}

\item{which_genome, }{genome assembly used ("hg19", "hg38")}

\item{qual_filt}{quality score filter above which is returned}
}
\value{
filtered bedpe list
}
\description{
Produce a list of bedpe, from a list of breakpoint GRanges,
   with score > qual_filt and both breakpoints within annotated genes
}
