% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_gridss.R
\name{gridss_parse_plot}
\alias{gridss_parse_plot}
\title{GRIDSS functions}
\usage{
gridss_parse_plot(vcf, dict_file, which_genome = NULL, output_path = NULL)
}
\arguments{
\item{vcf, }{the VCF file path}

\item{dict_file}{dictionary file for fasta used in alignment}

\item{which_genome, }{genome assembly used ("hg19", "hg38")}

\item{output_path, }{file path to output plots}
}
\value{
none, plots recurrent and private SVs and writes output to TSV format file
}
\description{
Wrapper to plot multi-sample VCF list (recurrent and not)
}
