% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{loh_summary}
\alias{loh_summary}
\title{LOH summary: find LOH in GRanges}
\usage{
loh_summary(join_chr_all_gr)
}
\arguments{
\item{join_chr_all_gr}{GRanges from master_mcols do.call cbind}
}
\value{
GRanges gr with LOH_sum,samples
}
\description{
LOH summary: find LOH in GRanges
}
