% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_pairtree.R
\name{make_pairtree_input}
\alias{make_pairtree_input}
\title{Write output for pairtree based on mutations in master TSV input}
\usage{
make_pairtree_input(
  rdata_input,
  cn_master,
  cn_pattern,
  pp_pattern,
  which_genome,
  tag
)
}
\arguments{
\item{rdata_input}{is RData from process vcfGra including master_all, and
gr_master_consensus_all}

\item{cn_master}{master CNA from fctscon process in Nextflow pipeline)}

\item{cn_pattern}{pattern matching FACETS input (fit_cncf_jointsegs.tsv)}

\item{pp_pattern}{pattern matching FACETS ploidy/purity (fit_ploidy_purity.tsv)}

\item{which_genome}{is GRCh37 or GRCh38}

\item{tag}{is string to tag output}
}
\value{
none, writes a tsv for input to pairtree
}
\description{
Format:
id,	name,	var_reads,	total_reads,	var_read_prob
id is sequential s0..sn
name is ID for the variant (chr:pos_ref>alt_HGVS)
the below are comma-sep
var_reads is reads supporting variant at variant locus
total_reads is all reads at variant locus
var_read_prob is based on CNA at locus; mutation at diploid locus is 0.5;
in a duplicated (n=3) locus, 1/3 etc.
}
