% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_pairtree.R
\name{make_pairtree_json}
\alias{make_pairtree_json}
\title{Write JSON for pairtree}
\usage{
make_pairtree_json(pyclone_res, pairtree_psm, tag)
}
\arguments{
\item{pyclone_res}{results from pyclone-vi write-results-file}

\item{pairtree_psm}{pairtree psm from make_pairtree_input(), with mut_name col}

\item{tag}{is string to tag output}
}
\value{
none, writes JSON for pairtree
}
\description{
Write JSON for pairtree
}
