% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R,
%   R/somenone_variantconsensus.R
\name{master_intersect_grlist}
\alias{master_intersect_grlist}
\title{Find intersect from list of CNA GRanges}
\usage{
master_intersect_grlist(gr_list, ps_vec, dp_vec, tag)

master_intersect_grlist(gr_list, ps_vec, dp_vec, tag)
}
\arguments{
\item{gr_list}{list of named GRanges objects}

\item{ps_vec}{mcols columns to keep p(er) s(ample; appended with list element's name)}

\item{dp_vec}{mcols columns to d(edu)p(licate); appended as-is if all same; if varaitions, usual 'dot number' formatting applies)}

\item{tag}{to apply to output file (master)}

\item{cna_list}{list of named GRanges objects}
}
\value{
GRanges object of all CNA, intersecting or not
        includes the ranges and each mcol from ps_vec, annotated `ps_vec[x].sample`

list of GRanges object of shared intersecting, and all mutations with ps_vec, dp_vec columns, and original rownames (should be unique therefore)
}
\description{
Find intersect from list of CNA GRanges

Find intersect from list of CNA GRanges
}
