% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{master_mcols}
\alias{master_mcols}
\title{Function to make master table mcols}
\usage{
master_mcols(gr_list, gr_master, ps_vec)
}
\arguments{
\item{gr_list}{same input as to master_intersect_cna_grlist()}

\item{gr_master}{object resulting from}

\item{ps_vec}{the genome assembly used, "hg19" or "hg38"}
}
\value{
list of master_df, data frame of master results
}
\description{
Function to make master table mcols
}
