% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someone_evoverse.R
\name{mut_cn_to_evoverse}
\alias{mut_cn_to_evoverse}
\title{Parse per-sample VCF and CN inputs, creating a list for evoverse input
evoverse is github.com/caravagna/mobster, cnaqc, viber, ctree etc.
Format:
chr, from, to, ref, alt, is_driver, driver_label, FILTER, DP, NV, VAF
ref/alt are the bases, DP total depth and NV No. reads at Variant, VAF = NV/DP
mut_rdata is master_consensus_shared from somatic_n-of-1 workflow
For somatic_n-of-1 workflow, CN is based on FACETS inputs
Format:
chr, from, to, Major, minor (not sic)
Purity also required as float}
\usage{
mut_cn_to_evoverse(mut_rdata, cn_pattern, pp_pattern, which_genome, tag)
}
\arguments{
\item{mut_rdata}{is RData file with shared mutations in 2+ samples
 [1] "seqnames"             "start"                "end"
 [4] "width"                "strand"               "rowname"
 [7] "Consequence"          "IMPACT"               "SYMBOL"
[10] "HGVSc"                "HGVSp"                "HGVSp1"
[13] "CLIN_SIG"             "samples_n"            "sampleIDs"
[16] <sampleIDn.AD>       "sampleIDn.AD.1"     "sampleIDn.AF"}

\item{cn_pattern}{is a pattern that matches FACETS input filenames}

\item{pp_pattern}{is a pattern that matches polidy/purity table from FACETS}

\item{which_genome}{is GRCh37 or GRCh38}
}
\value{
none, writes a tsv for input to Chimaera
}
\description{
Parse per-sample VCF and CN inputs, creating a list for evoverse input
evoverse is github.com/caravagna/mobster, cnaqc, viber, ctree etc.
Format:
chr, from, to, ref, alt, is_driver, driver_label, FILTER, DP, NV, VAF
ref/alt are the bases, DP total depth and NV No. reads at Variant, VAF = NV/DP
mut_rdata is master_consensus_shared from somatic_n-of-1 workflow
For somatic_n-of-1 workflow, CN is based on FACETS inputs
Format:
chr, from, to, Major, minor (not sic)
Purity also required as float
}
