% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_pairtree.R
\name{mut_cn_to_pyclone}
\alias{mut_cn_to_pyclone}
\title{Parse mutation and CN input, creating a master table of PyClone(VI) input
Format:
mutation_id, sample_id, ref_counts, alt_counts, normal_cn, major_cn, minor_cn, tumour_content
mutation_id is a unique ID (position-based)
sample_id is sample
ref/alt_counts are read counts at Mutation, VAF the variant allele frequency
mut_rdata is master_consensus_shared from somatic_n-of-1 workflow
For somatic_n-of-1 workflow, CN is based on FACETS inputs}
\usage{
mut_cn_to_pyclone(mut_gr, cn_pattern, pp_pattern, tag)
}
\arguments{
\item{mut_gr}{gr_master_consensus_all}

\item{cn_pattern}{pattern matching FACETS input (fit_cncf_jointsegs.tsv)}

\item{pp_pattern}{pattern matching FACETS polidy/purity (fit_ploidy_purity.tsv)}
}
\value{
none, writes a tsv for input to pyclone(VI)
}
\description{
Parse mutation and CN input, creating a master table of PyClone(VI) input
Format:
mutation_id, sample_id, ref_counts, alt_counts, normal_cn, major_cn, minor_cn, tumour_content
mutation_id is a unique ID (position-based)
sample_id is sample
ref/alt_counts are read counts at Mutation, VAF the variant allele frequency
mut_rdata is master_consensus_shared from somatic_n-of-1 workflow
For somatic_n-of-1 workflow, CN is based on FACETS inputs
}
