% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{output_out_list}
\alias{output_out_list}
\title{Plotting and Output from 'out_list' object}
\usage{
output_out_list(out_list, in_list, dict_file, which_genome, tag, cgc_gr = NULL)
}
\arguments{
\item{out_list}{resulting from process_in_list}

\item{in_list}{input to process_in_list}

\item{dict_file}{dictionary file of fasta used to align BAMs}

\item{which_genome}{the genome assembly used, "hg19" or "hg38"}

\item{tag}{runID and annotation strategy, "ENS" or "CGC"}
}
\value{
none writes output files
}
\description{
Plotting and Output from 'out_list' object
}
