% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_pairtree.R
\name{parse_cn_list}
\alias{parse_cn_list}
\title{Parse CN files and output as list}
\usage{
parse_cn_list(mut_gr, cn_pattern, pp_pattern)
}
\arguments{
\item{mut_gr}{is the GRanges of Mutation data}

\item{cn_pattern}{is a pattern that matches CN input filenames}

\item{pp_pattern}{is a pattern that matches polidy/purity table from FACETS}
}
\value{
list of correct format per sample
}
\description{
Parse CN files and output as list
}
