% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someone_evoverse.R
\name{parse_pur_list}
\alias{parse_pur_list}
\title{Parse purity files and output as list}
\usage{
parse_pur_list(pp_patern)
}
\arguments{
\item{pp_pattern}{is a pattern that matches polidy/purity table from FACETS}
}
\value{
list of correct format per sample
}
\description{
Parse purity files and output as list
}
