% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someone_evoverse.R
\name{parse_sample_cn_list}
\alias{parse_sample_cn_list}
\title{Parse CN files and output as list}
\usage{
parse_sample_cn_list(cn_pattern)
}
\arguments{
\item{cn_pattern}{is a pattern that matches CN input filenames}
}
\value{
list of correct format per sample CNs
        chr, from, to, Major, minor (not sic)
}
\description{
Parse CN files and output as list
}
