% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someone_evoverse.R
\name{parse_sample_mut_list}
\alias{parse_sample_mut_list}
\title{Parse mut_df and output as list per sample}
\usage{
parse_sample_mut_list(sampleIDs, mut_df, vaf_filter = 0.05)
}
\arguments{
\item{mut_df}{is data frame of data with samples <sampleID[1].AD, .AD.1, .AF}

\item{vaf_filter}{is level of VAF to filter, default 5%}

\item{sampleIDS}{is vector of sample IDs to parse out depth data}
}
\value{
list of correct format per sample
}
\description{
Parse mut_df and output as list per sample
}
