% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{plot_cna_df}
\alias{plot_cna_df}
\title{Processing list of Facets input}
\usage{
plot_cna_df(cna_df, cum_sum_add, cna_max = 8)
}
\arguments{
\item{in_list}{of data from Facets}
}
\value{
purity-ploidy dataframe, also write rds used
}
\description{
Processing list of Facets input
}
