% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Create plot of shared variants among samples}
\usage{
plot_consensus(
  master_gr,
  tag,
  included_order,
  sample_map = NULL,
  colours = NULL,
  plot_label_pattern = "patho"
)
}
\arguments{
\item{tag}{is a string to tag output files}

\item{included_order}{ordering of samples for plotting}

\item{sample_map}{map included_order to new names, must be name vector where
names equate to included_order elements}

\item{colours}{to use for colouring/shading, made into a rampPalette}

\item{plot_label_pattern}{match this to print label of variant
(NB too vague and you will have a huge amount of labels which looks shit;
currently set to show 'patho'genic)}

\item{plot_list}{is a nested list of plot data [[caller]][[samples1..n]]}

\item{raw_list}{is a nested list of raw calls [[caller]][[samples1..n]]}
}
\value{
none, plots PDF and writes out tsv files
}
\description{
Create plot of shared variants among samples
}
