% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{plot_consensus_single}
\alias{plot_consensus_single}
\title{Create two plots: all consensus, those in 2+ samples}
\usage{
plot_consensus_single(plot_list, raw_list, tag, included_order)
}
\arguments{
\item{plot_list}{is a nested list of plot data [[caller]][[samples1..n]]}

\item{raw_list}{is a nested list of raw calls [[caller]][[samples1..n]]}

\item{tag}{is a string to tag output files}
}
\value{
data.frame object of raw variant call allele frequencies
}
\description{
Create two plots: all consensus, those in 2+ samples
}
