% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_gridss.R
\name{prep_plot_circos_sv}
\alias{prep_plot_circos_sv}
\title{OmicCircos plotting from data.frame from GRIDSS bedpe}
\usage{
prep_plot_circos_sv(input_df, which_genome, dict_file, output_path)
}
\arguments{
\item{input_df}{data.frame with chrom1, start1, end1, symbol1,
chrom2, start2, end2, symbol2, sampleID, colour columns}

\item{which_genome, }{genome assembly used ("hg19", "hg38")}

\item{dict_file}{dictionary file for fasta used in alignment}

\item{output_path}{path to where PDF file with output plot is written
N.B. two pages, one with Circos plot and one with legend}
}
\value{
none, prints plot to file, and a file too
}
\description{
OmicCircos plotting from data.frame from GRIDSS bedpe
}
