% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{seqlengths_df}
\alias{seqlengths_df}
\title{Create lengths of chromosomes for plotting}
\usage{
seqlengths_df(in_seqs, dict_file, which_genome)
}
\arguments{
\item{in_seqs}{a jointsegs filename (output from Facets) with columns:
"seg", "num.mark", "nhet", "cnlr.median", "mafR", "segclust", "cnlr.median.clust", "mafR.clust", "cf.em", "tcn.em", "lcn.em"}

\item{dict_file}{dictionary file for fasta used in alignment}

\item{which_genome}{the genome assembly used, "hg19" or "hg38"}
}
\value{
a GRanges object with annotated jointsegs from input
}
\description{
Create lengths of chromosomes for plotting
}
