% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{sub_hgvsp}
\alias{sub_hgvsp}
\title{Create single-letter HGVS protein annotation (VEP outputs 3-letter)
   make vector, gsub out aa3 for aa1}
\usage{
sub_hgvsp(in_vec)
}
\arguments{
\item{in_vec}{vector input}
}
\value{
vector of single-letter HGVS protein IDs
}
\description{
Create single-letter HGVS protein annotation (VEP outputs 3-letter)
   make vector, gsub out aa3 for aa1
}
