% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{variant_consensus}
\alias{variant_consensus}
\title{somtic_variant_consensus functions}
\usage{
variant_consensus(
  germline_id,
  vep_vcf_pattern,
  raw_vcf_pattern = "raw.vcf",
  tag = "somatic_n_of_1",
  included_order = NULL,
  name_callers = NULL
)
}
\arguments{
\item{germline_id}{ID for germline_id sample}

\item{vep_vcf_pattern}{pattern to match VEP annotated VCFs}

\item{raw_vcf_pattern}{pattern to match raw unannotated, unfiltered VCFs unfiltered}

\item{tag}{is a string to tag output files}

\item{included_order}{oredering of samples for plotting}

\item{name_callers}{named variant callers to use primary to random selection}
}
\value{
vector of single-letter HGVS protein IDs
}
\description{
Primary function to call others and produce plots, table
}
