% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{vcf_parse_gr}
\alias{vcf_parse_gr}
\title{Parses for VCFs into GRanges object}
\usage{
vcf_parse_gr(vcf_in, germline_id)
}
\arguments{
\item{germline_id}{ID for germline_id sample}

\item{in_vec}{vector input}
}
\value{
vector of single-letter HGVS protein IDs
}
\description{
Parses for VCFs into GRanges object
}
