% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{vcf_vep_ann_parse_soma_gr}
\alias{vcf_vep_ann_parse_soma_gr}
\title{Parses for VCFs annotated by VEP into GRanges object
   takes CANONICAL annotation or first if no CANONICAL}
\usage{
vcf_vep_ann_parse_soma_gr(vcf_in, germline_id)
}
\arguments{
\item{germline_id}{ID for germline_id sample}

\item{in_vec}{vector input}
}
\value{
vector of single-letter HGVS protein IDs
}
\description{
Parses for VCFs annotated by VEP into GRanges object
   takes CANONICAL annotation or first if no CANONICAL
}
