% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{write_consensus_all}
\alias{write_consensus_all}
\title{Return overlapping variants for all samples}
\usage{
write_consensus_all(plot_list, plot_df, tag, included_order, cons)
}
\arguments{
\item{plot_list}{list produced by at_least_two()}

\item{plot_df}{data.frame of plotting information}

\item{tag}{is a string used to tag output files}

\item{included_order}{ordering of samples for plotting}

\item{cons}{string to define consensus (shared or all)}
}
\value{
GRanges object of all  of single-letter HGVS protein IDs
}
\description{
Return overlapping variants for all samples
}
