% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{facets_cna_consensus}
\alias{facets_cna_consensus}
\title{Consensus CNA estimation and plotting with Facets input}
\usage{
facets_cna_consensus(pattern, dict_file, tag, cgc_bed = NULL)
}
\arguments{
\item{pattern}{a pattern within file names to use for parsing input}

\item{dict_file}{a dictionary file for the fasta used to align data}

\item{tag}{a string used to tag output}

\item{cgc_bed}{bedfile from the Cancer Gene Census from COSMIC}
}
\value{
none, saves plots and data in current dir
}
\description{
Consensus CNA estimation and plotting with Facets input
}
