% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{facets_jointsegs_parse_to_gr}
\alias{facets_jointsegs_parse_to_gr}
\title{Parse jointsegs into GRanges object}
\usage{
facets_jointsegs_parse_to_gr(
  jointsegs_in,
  which_genome,
  cgc_gr = NULL,
  anno = NULL,
  bsgenome = NULL
)
}
\arguments{
\item{a}{string either "hg19" or "hg38"}
}
\value{
a GRanges object with annotated jointsegs from input
}
\description{
Parse jointsegs into GRanges object
}
