% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_gridss.R
\name{plot_circos_sv}
\alias{plot_circos_sv}
\title{OmicCircos plotting from data.frame from GRIDSS bedpe}
\usage{
plot_circos_sv(input_df, output_path, cytoband)
}
\arguments{
\item{input_df}{data.frame with chrom1, start1, end1, symbol1,
chrom2, start2, end2, symbol2, sampleID, colour columns}

\item{output_path}{path to where PDF file with output plot is written
N.B. two pages, one with Circos plot and one with legend}

\item{cytoband}{internal file downloaded in prep_plot_circos_sv}
}
\value{
none, prints plot to file
}
\description{
OmicCircos plotting from data.frame from GRIDSS bedpe
}
