% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_facets.R
\name{process_in_list}
\alias{process_in_list}
\title{Processing list of Facets input}
\usage{
process_in_list(in_list, which_genome, cgc_gr)
}
\arguments{
\item{in_list}{of data from Facets}

\item{which_genome}{genoem assembly being used "hg19" or "hg38"}

\item{cgc_gr}{GRanges of cancer gene census from COSMIC}
}
\value{
purity-ploidy dataframe, also write rds used
}
\description{
Processing list of Facets input
}
