% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somenone_variantconsensus.R
\name{write_consensus_all}
\alias{write_consensus_all}
\title{Return overlapping variants for all samples
write VCF for PCGR based on example BRCA VCF from Github}
\usage{
write_consensus_all(plot_list, plot_df, tag, included_order, cons)
}
\arguments{
\item{plot_list}{list produced by at_least_two()}

\item{plot_df}{data.frame of plotting information}

\item{tag}{is a string used to tag output files}

\item{samp}{is the sample being operated on}
}
\value{
GRanges object of all  of single-letter HGVS protein IDs
}
\description{
Return overlapping variants for all samples
write VCF for PCGR based on example BRCA VCF from Github
}
